custom-elements/instantiating-custom-elements/non-configurable-constructor-property.html

<!DOCTYPE html>
<html>
<head>
<title>PROTOTYPE has a non-configurable property named constructor</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If  PROTOTYPE has a non-configurable property named constructor, throw a NotSupportedError and stop">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    var proto = Object.create(HTMLElement.prototype);
    Object.defineProperty(proto, 'constructor', {configurable: false});
    //FIXME Should the type of the exception be specified here?
    assert_throws(null, function() {
        doc.registerElement('x-a-1', {prototype: proto});
    }, 'Exception should be thrown in case of attempt to register element ' +
        'with a non-configurable property named constructor');

}, 'Prototype has a non-configurable property named constructor');


test(function() {
    var doc = newHTMLDocument();
    
    var proto = Object.create(HTMLElement.prototype);
    Object.defineProperty(proto, 'constructor', {configurable: true});
    
    try {
        doc.registerElement('x-a-2', {prototype: proto});
    } catch (e) {
        assert_unreached('Exception should not be thrown in case of attempt to register ' +
        'element with a configurable property named constructor');
    }
}, 'Prototype has a configurable property named constructor');

</script>
</body>
</html>
